% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{automatedReadabilityIndex}
\alias{automatedReadabilityIndex}
\title{Calculates the automated readability index from a given text.}
\usage{
automatedReadabilityIndex(text)
}
\arguments{
\item{text}{An input character string.}
}
\description{
Calculates the automated readability index from a given text.
}

