% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{chunk}
\alias{chunk}
\title{Utility functions}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{A vector to be splitted.}

\item{n}{An integer specifying the number of chunks.}
}
\description{
Splits a sequence into equal parts.
}

