% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{ClassificationDiagnostics}
\alias{ClassificationDiagnostics}
\title{Provides classification diagnostics for a combination of predicted and reference values.}
\usage{
ClassificationDiagnostics(prediction, trueValues)
}
\arguments{
\item{trueValues}{A numeric vector.}

\item{predicted}{A numeric vector.}
}
\description{
Provides classification diagnostics for a combination of predicted and reference values.
}

