% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{GoodnessOfFit}
\alias{GoodnessOfFit}
\title{Performance metrics}
\usage{
GoodnessOfFit(predicted, trueValues, p = NULL)
}
\arguments{
\item{predicted}{A numeric vector.}

\item{trueValues}{A numeric vector.}

\item{p}{The number of predictor variables in the model. Required to calculate adjusted R^2 (optional)}
}
\description{
Calculates goodness-of-fit for a combination of predicted and reference values.
}

