% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{labeledVector}
\alias{labeledVector}
\title{Produces a vector of a desired length with an arbitrary number of labels.
Each label occurs equally frequent in the resulting sequence.}
\usage{
labeledVector(vectorLength, numberOfChunks)
}
\arguments{
\item{vectorLength}{An integer specifying the length of the vector.}

\item{numberOfChunks}{An integer specifying the number of labels}
}
\description{
Produces a vector of a desired length with an arbitrary number of labels.
Each label occurs equally frequent in the resulting sequence.
}

