% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm_stepwise}
\alias{lm_stepwise}
\title{Performs simple stepwise regression.}
\usage{
lm_stepwise(dependent, independent, data, controlVariables = NULL)
}
\arguments{
\item{dependent}{Name of the dependent variable.}

\item{independent}{A vector containing the names of the independent variables.}

\item{data}{A data frame.}

\item{controlVariables}{A vector containing the names of the control variables. Default is \code{NULL}.}
}
\description{
Performs simple stepwise regression.
}

