% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{loadEnglishDictionary}
\alias{loadEnglishDictionary}
\title{Load MIT English dictionary}
\usage{
loadEnglishDictionary(stemDictionary = TRUE)
}
\arguments{
\item{stemDictionary}{If \code{TRUE}, each word is stemmed to its root form (default).}
}
\value{
Vector of valid English words
}
\description{
Load MIT English dictionary
}

