% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{loadHarvardDictionary}
\alias{loadHarvardDictionary}
\title{Text mining functions}
\usage{
loadHarvardDictionary(stemDictionary = TRUE)
}
\arguments{
\item{stemDictionary}{If \code{TRUE}, each word is stemmed to its root form (default).}
}
\value{
list of positive and negative stems
}
\description{
Load Harvard IV psychological dictionary
}

