% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{loadHenryDictionary}
\alias{loadHenryDictionary}
\title{Load Henry dictionary}
\usage{
loadHenryDictionary(stemDictionary = TRUE)
}
\arguments{
\item{stemDictionary}{If \code{TRUE}, each word is stemmed to its root form (default).}
}
\value{
list of positive and negative stems
}
\description{
Load Henry dictionary
}

