% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{loadLMDictionary}
\alias{loadLMDictionary}
\title{Load LM dictionary}
\usage{
loadLMDictionary(stemDictionary = TRUE)
}
\arguments{
\item{stemDictionary}{If \code{TRUE}, each word is stemmed to its root form (default).}
}
\value{
list of positive and negative stems
}
\description{
Load LM dictionary
}

