% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{loadSentiWordNetDictionary}
\alias{loadSentiWordNetDictionary}
\title{Load SentiWordNet dictionary}
\usage{
loadSentiWordNetDictionary(stemDictionary = TRUE)
}
\arguments{
\item{stemDictionary}{If \code{TRUE}, each word is stemmed to its root form (default).}
}
\value{
list of positive and negative stems
}
\description{
Load SentiWordNet dictionary
}
\note{
Numeric terms and terms with neutral sentiment are excluded.
}

