% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{polarity}
\alias{polarity}
\title{Produces polarity labels for a numeric input vector.}
\usage{
polarity(x, labelPos = 1, labelNeg = -1, labelNeutral = NA)
}
\arguments{
\item{x}{A numeric input vector.}
}
\description{
Produces polarity labels for a numeric input vector.
}

