% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize a string vector.}
\usage{
tokenize(x)
}
\arguments{
\item{x}{A vector to be tokenized.}
}
\value{
Tokenized vector.
}
\description{
Tokenize a string vector.
}

