% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{variance_inflation_factors}
\alias{variance_inflation_factors}
\title{Calculates variance inflation factors for dataframe and response variable.}
\usage{
variance_inflation_factors(x, y)
}
\arguments{
\item{x}{A dataframe.}

\item{y}{A numeric response variable.}
}
\value{
Vector with variance inflation factors for all variables in x.
}
\description{
Calculates variance inflation factors for dataframe and response variable.
}

