% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{winsorize}
\alias{winsorize}
\title{Winsorizes a given numeric input vector.}
\usage{
winsorize(x, fraction = 0.05)
}
\arguments{
\item{x}{A numeric vector.}

\item{fraction}{Fraction of values to be winsorized. Default is 0.05.}
}
\description{
Winsorizes a given numeric input vector.
}

